requirements.txt
cat > requirements.txt << 'EOF'
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.3
requests==2.31.0
EOF